Refactor file paths and names
